requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
duckdb>=0.9.2
folium>=0.14.0
streamlit-folium>=0.14.0
altair>=5.0.0
pandas>=2.0.0
polyline>=1.4.0
stravalib>=1.5.0
requests>=2.31.0
